raft_server: redesign the proto to allow the background timer to work together with the rpc server.
- Initially I considered two designs (refer to
https://docs.google.com/drawings/d/1lG4Oza8qSlVZeBheTK7pXb--bCNR3MAz0OGg_HQxLzE/edit?usp=sharing), both of them are seemed to
have good performance. In the end I chose the one more easy to implement (B in the drawing).
Briefly, the difference between design A and design B is the way the timer notify the FSM to tick(RawNode::Tick).
In desgin B, the rpc server runs with a background timer, which sends TickRequest to server at a fixed interval. The TickRequest is an
IPC. The design A communicates between threads using the dispatcher, without using IPC.
Apparently design B is more easy to implement, and the TickRequest IPC is not so frequent, that doesn't seem to affect the performance.