rpc: fix bug that background timer sends tick to rpc server when it's not set up.
We encountered a bug when we setup a node configured with 5 peers, and the node
crashed every time when it starts the background timer.

Though we still unaware of the cause, we found that background timer will sends
tick request to the RPC server even though it has not set up. Maybe this is the
 real cause for the node crash.