src: global ready flusher
In this commit, we optimized the original per-ReplicatedLog-instance-per-ReadyFlusher-thread
design, and now, we only need to start a single global ReadyFlusher for all
of the ReplicateLog instances.

Our implementation is still naive. Flushing tasks are managed by the thread pool
of std::async. ReadyFlusher only begin another round of flushing (which means
a for-loop flushing all replicas) after it ends the previous one.

Another choice is to wait for a specified number of time after the for-loop ends,
and it doesn't have to stall on the previous round.